Added a Second PostImage Component for RSS
The Next.js Image component was causing issues when generating the HTML from the markdown in the RSS feeds causing a 2nd “blank” image to appear. I guess browsers can support that better. I added a second `PostImage` component that takes the same props but will use the standard `<img/>` tag. This component is loaded in the MDX renderer for RSS to prevent the duplicate image.

Fixes #316